+++ internal/manifests/manifestutils/paths.go
@@ -5,4 +5,6 @@ const (
 	TLSDir = "/var/run/tls"
 	// CABundleDir is the path that is mounted from the configmap for TLS.
 	CABundleDir = "/var/run/ca"
+	// CAReceiver is the path that is mounted from the configmap for TLS for receiver.
+	CAReceiver = "/var/run/ca-receiver"
 )